[Reminders] Refactor remind module

+++ TheGodfather/Modules/Owner/Services/PrivilegedUserService.cs
@@ -13,9 +13,10 @@ namespace TheGodfather.Modules.Owner.Services
         public PrivilegedUserService(DbContextBuilder dbb)
             : base(dbb) { }
 
+
         public override DbSet<PrivilegedUser> DbSetSelector(TheGodfatherDbContext db) => db.PrivilegedUsers;
         public override PrivilegedUser EntityFactory(ulong id) => new PrivilegedUser { UserId = id };
         public override ulong EntityIdSelector(PrivilegedUser entity) => entity.UserId;
-        public override object[] EntityPrimaryKeySelector(ulong id) => new object[] { id };
+        public override object[] EntityPrimaryKeySelector(ulong id) => new object[] { (long)id };
     }
 }
